+++ NEWS.md
@@ -1,6 +1,6 @@
 # CHANGES IN bookdown VERSION 0.26
 
-- Added support for theorem/proof environments back for Word/EPUB output (thanks, @N0rbert, #1313).
+- Added support for theorem/proof environments back for Word/EPUB/ODT output (thanks, @N0rbert, #1313).
 
 - `kindlegen()` is defunct now.